+++ examples/base/src/layouts/UserLayout.tsx
@@ -21,7 +21,11 @@ const UserLayout = () => {
       ]}
       api="/api/login"
       btnProps={{ type: 'primary' }}
-      onBefore={() => ({ a: 12, b: 1221 })}
+      saveField={{
+        userName: 'username',
+        passWord: 'password',
+      }}
+      onBefore={(store) => ({ a: 12, b: 1221, ...store })}
       onSuccess={(data) => {
         if (data && data.token) {
           sessionStorage.setItem('token', data.token)